Move url redaction to its own file
